Sixth Demo
Now I'm using a static build of ffprobe, and an explicit library on npm to interact with the JSON objects it produces. This kind of could help conceptually with making things work completely in Rust too. Not sure why I'd do that for a small thing like this, probably won't, but you could. Say like I were to refactor this into Rust, it could omit Node.js and it could be a single executable. I want to try to design my projects onwards to be easy to do in that regard. Sometimes it might not be as feasible or possible, but it might be good to work towards that.