Concurrency Test
Prompt:

How can I make this more performant, so it can run some of these calculations in parallel? I don't want to do all of them at once though, I don't want it to lock up my computer. There could be 30 songs it does, or up to 2000+, so some kind of threading is nice, but not at the expense of freezing everything.

Still some issues though, working on this now:

One issue though is that it no longer is able to log things out as they are calculated, and the order that they come in at is different than they were originally organized.